some refactoring, and scrypt interface has to be provided by caller, scryptsy most likely not working for RN
